requirements.txt
astroid==1.6.5
colorama==0.3.9
Django==2.0.6
djangorestframework==3.8.2
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.9.2
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
wrapt==1.10.11
